+++ NEWS.md
@@ -8,8 +8,9 @@ The inputs with tracking enabled are marked with a double arrow icon.
 The affected modules are: `tm_t_summary`, `tm_t_summary_by`, `tm_g_forest_rsp`, `tm_g_forest_tte`, 
 `tm_t_events_summary`, `tm_t_abnormality`, `tm_t_mult_events`.
 * Added an additional validation for `tm_g_patient_timeline` when the plot is empty.
-* Fixed Get R Code output of `rm_t_pp_laboratory` to return identical HTML 
+* Fixed Get R Code output of `tm_t_pp_laboratory` to return identical HTML 
 formatted table as displayed in the app.
+* Added the option to choose the number of decimal places for rounding in `tm_t_pp_laboratory`.
 
 # teal.modules.clinical 0.8.8
 ### Enhancements